Put generic db-generation code in blastDB (it should be in the rnacentral repo); added ran16s-related code with a simple predicate
